7179 FIX Fixed site filtering when creating availability PDF reports
When having a filtered availability view which shows information of specific
sites in distributed setups, the PDF reports created with "Export as PDF" were
not applying the site filter. In this situation a query was sent to all
distributed sites for fetching the data. It was not only querying too many
sites, the PDFs also contained more hosts / services than intended in this
situation.

Change-Id: Ia15a5533e4e45b27c97f7f0bd00322b4608c0f9b