7181 FIX Improved handling of common issues during distributed site login
When connecting a new distributed site to a central site an issue that could
happend accidentally was that the remote site had a pending change in WATO in
the moment the central site is establishing the configuration connection.

This resulted in an error message like this during the first activation:
"There are 1 pending changes that would get lost".

Since the user already confirms that the site configuration should be overwritten
during site login, the step now automatically confirms all pending changes.

Additionally the site login now creates a pending change for the new remote
site in the central WATO to trigger an activation for this site after the login
even without a configuration change.

Change-Id: I580e6f06f5b343fcc6867608a5e4a6829d9b6f3b