New generic item input validation method
When a dialog takes a parameter from HTTP variables that represents an
item of a collection, it normally searches for the item before
proceeding and raises an error in case it does not exist in the
collection.

This is done individually in many places with individual error messages
that all have to be localized individually. This new function does this
in a generic way and should reduce page individual code a bit.

Change-Id: I4b515f5d9c86a4e233d646fe3a22b4ad498d85e8