6762 FIX zfsget: Aligned behaviour in Solaris and Linux
Previously the implementation of getting zfs information was slightly
different in Linux and Solaris. Now both agents are fetching data only for
volumes and filesystems and are using the quota for computing the total size of
a filesystem or volume. So the information of each service will be more precise.

FEED-3375

Change-Id: I22d5f57c62a028af56b1bdf52b58e119b62f4588