7185 Custom service attributes can now be configured
It is now possible to configure arbitrary custom attributes for all services
just like it was already possible for hosts and users in previous versions.

The configuration procedure is working equally, you first need to define an
attribute in the first place, to be able to refer to the attribute and assign
it to services.  The definition of custom attributes is done using the Global
Setting "Custom service attributes".

Once you have defined a custom service attribute, you can assign it to a
collection of services using the equal named ruleset "Custom service attributes".

In general, you should keep the number of rules low. To support you in this,
the rule is structured so that you can select several custom attributes for
each rule.

CMK-1727

Change-Id: I9e8f21b04f4f5909b0f1881238864cd8dfa3a514