7190 More GUI elements can be shared with specific contact groups
Views, reports and dashboard could already be shared with the users of specific
contact groups.

For a consistent user experience other customizable elements can now also be
shared with a specific group of users instead of all users.

This change affects the following elemets: The bookmark lists and snapins for
all Check_MK Editions and the Graph Collections, Custom Graphs, Graph Tunings
and SLAs for the Enterprise and Managed Services Edition

CMK-1754

Change-Id: I32a0ebdd958e99d4b8628cf6ab24518bdc693f7b