Start moving view logic into dedicated object
The goal is to restructure the view processing logic classes
that separate the logic and make the whole view rendering
easier to understand.

First steps: Move logic from page_view() and show_view() to
the new View() class to be able to clean up some hacks related
to availability and SLA processing.

Change-Id: Ic1da2b56dd361cce33f1a02d867c2681d4a0f197