7062 Improvement of the graphing with PNP4Nagios for various checks
When PNP4Nagios is used Check_MK tries to use autogenerated graph templates
to paint the graphs of a check. If the capabilities for Check_MKs autogeneration
of graph templates is insufficient Check_MK provides a fallback to handcrafted
templates.

This werk extends the capabilities of Check_MKs autogeneration of PNP templates
and makes a lot of handcrafted PNP templates unnecessary. Therefore, these
templates are removed. This results in an overall more pleasant and uniform
graphing.

PNP4Nagios is used if the raw edition is used or if the enterprise edition is
used together with the Nagios core.

Change-Id: I31c680f89d307e054ff2428b2065a01de484c555