7228 FIX Fixed failing liveproxyd activation when process is currently starting
There was a race condition in the code that made configuration changes fail
with a message like "Could not reload Livestatus Proxy: invalid literal for
int() with base 10: \'\'. See web.log for further information." when one
tried to chang the site configuration while the proxy was currently in a
specific part of the process startup phase.

Change-Id: I88de20d46895643bf8b811493ccfc4f9af97eabd