7155 Ended Support for SLES 11 in version 1.6
Since SLES 11 are using a OpenSSL version before 1.0.0 we decided to end the support for SLES 11 for our next release. Many Python packages, that are used in Check_MK are dependent on newer OpenSSL versions. An old OpenSSL version is not only a security risk but also a threat for compatibility and stability.

Change-Id: I4bbd323c99ceaa78e389f34980df8c727a858611