7230 FIX Fixed selection of checks in "Check intervals for SNMP checks" ruleset
Instead of selecting individual check types as it was possible before, the section names
are now selectable as it was intended. As described in help text of this option it is
not possible to configured this setting for individual sub checks, but only for a family
of SNMP checks that share a single SNMP section.

FEED-3402

Change-Id: Ie1948a4478451a9961a06fb0503a7d4e229db73a