if.include: rigorously make if-scans mutually exclusive
7214 Interface checks: Introduce scan function API

In the past conflicting check plugins with services called "Interface <ITEM>"
could lead to a "duplicate services" warning on the discovery page.

To avoid this issue in the future, all scan functions for those plugins
are now located in the file "if.include".
If you add a new "Interface <ITEM>"- service, please add the corresponding
scan function to the file "if.include", and list in the dictionary
<tt>_SPECIFIC_IF_SCANS</tt>. You can then access the scan function by specifying
<tt> 'snmp_scan_function': get_snmp_if_scan_function('my_if_check')</tt>
in your check plugin.
The matching devices will then be automatically excluded from the generic if
and if64 check plugin.

7215 FIX Avoid duplicate services for if_fortigate check

This fixes the duplicate services warning displayed on the
discovery page for if/if_fortigate checks.

Services of the if checkplugin are no longer discovered if
the if_fortigate plugin already found interfaces.

Change-Id: Id6736a24fcce8f235354213bd3708527ea5b8779