7065 FIX k8s_stats: fix crashes regarding timestamp parsing
The Kubernetes API provides timestamps for the k8s_stats checks
in the RFC3339Nano format with different timezones and a varying
number of trailing zeros in the seconds field. This lead to
a crash of the k8s_stats.network check. To resolve this issues
the timestamps are now converted to unix timestamps in the
special agent.

Change-Id: I2869540e2b6b73b0071379c307b4e2615f6235f4