7234 FIX Fixed broken event cancelling by priority (1.5.0p12 regression)
When a EC rule is configured to cancel an open event by priority and the lower
priority level is the left of the priority range, the cancelling was not
working.

For example EMERG to CRIT worked while CRIT to EMERG did not work. This has
been fixed. The order of both levels should not be relevant for the processing
of the rules.

Change-Id: I7a285338fb1c716b50503c37fe0b3b55a27fd364