7235 WATO now stores host/folder creation meta information
When either a host or folder is being created in WATO it now stores the
creation time and creator ina meta data attribute. This attribute is shown at
the bottom of the the host or folder edit dialog.

Hosts or folders created before 1.6 don't have such information stored and will
show something like "Created sometime before 1.6".

Change-Id: I9c26dc71dfdc50da5f98a67297494e38d299272a