7236 Deterministic rules.mk when "Use GIT version control for WATO" is enabled
When "Use GIT version control for WATO" is enabled to keep track of the changes
that have been made to the Check_MK configuration, the history of the rules.mk
files was problematic because several contents of that file were not saved in a
deterministic way. This means that loading and saving the configuration without
chaning anything resulted in a changed file, e.g. because attributes of
dictionaries were not sorted.

The rules.mk is now stored in a deterministic way when "Use GIT version control for WATO"
is enabled.

Change-Id: Ica1cf3e63e170ba4f61fbe521c9bb667070ad0bd