6658 FIX Differentiation of host & service problems notifications
For REST notification plugins like Slack, PagerDuty and VictorOPS. Problems
where classified by the presence of the SERVICESTATE context variable. But
in the RAW edition this variable is completed by its macro and thus the
classification leads to host problems to be described as service problems
too. Classifying after the WHAT context variable solves the issue.

Change-Id: I08aa651604c60d2810777c73746ba43ede14c2f3