Fix inefficient event history filtering by event ID
Queries for specific event IDs were not optimized for the event history
queries which are issues e.g. by the GUI when clicking on the detail
page of one event.

This is only the first part of the fix, the second part will be added in
a separate commit for the livestatus code.

Change-Id: I991fa8d9c1c4d06316f7a4fb1b03e99aedd6c204