7242 FIX Unreachable site errors respect site filters
When reports are being created while some distributed site is offline, the
report is prefixed by a "error report" page that is added to inform the user
that the report was created in such a situation and some data may be missing
for this reason.

The error message looks like this: "Not all sites have been queried. Some
information is missing."

The errors were shown for all sites, even when a specific site was queried.
This was a problem, especially in Check_MK Managed Services Edition
environments when creating customer specific reports in the management sites.

Change-Id: Ice1b70be3c1aeef2d689f50f5e6ae2575794688d