7068 FIX Truncate values of the notification context which are too long
If the the value stored in the notification context was larger than
the OS limit MAX_ARG_STRLEN a failed notification was produced and
the error 'OSError: [Errno 7] Argument list too long' appeared in
the notify.log. This was usually the case for very long values in
the long outputs of a service like e.g. fileinfo.groups. Now the
values are truncated and the text 'Attention: Removed remaining
content because it was too long.' is shown at the end.

Change-Id: I726dc536378f54da5dccf624d551b1565ecf0d47