ps.include: remove break statement
Unfortunately, we have to check all lines against all specs:
Even if a line already matched an earlier spec, it may match
another spec as well (and create a new item).
Even if a spec found a matching line earlier, it may match a
different line and create a different item when the matching
group changes.

Change-Id: Iab8be87adc2f728283ac40b353f233ab94f52b43