7222 FIX check_http: port config for certificate check via proxy
If users had the active check "Check HTTP" configured to check a certificate
via a proxy, the option "TCP Port" had wrongly been used as the proxys port
instead of the certificate servers port.

The port defined in the option "TCP Port" is now applied to the certificate
server, regardless of whether a proxy is used.

If you relied on the wrong behaviour, remove the option "TCP Port" and include
the port for the proxy in the proxy settings.

NOTE:
To make this work, we have to pass proxy address and server name not
using the "-I proxy.com" and "-H server.com" option, but as the first
and second argument, respectively. This weird hack bypasses a bug in
the check_http commandline parsing:
When we pass the arguments this way, we can include the *servers* port
in the second argument (e.g. "server.com:8443"). When we try the same
with "-H server.vom:8443" the port will be stripped, and ignored if
the "-p" option is set, and used as the proxys port otherwise.

Change-Id: I2f7babed02b652186f996a91e53048a3689c9ebb