Cleanup: Synchronous Rule.to_dict_config and Rule.parse_dict_rule()
Both function now deal with optional and mandatory attributes in the
same way.

These dicts will be used as rule representation in the configuration
files in the future. Therefore we need to skip the attributes that are at
their default values to reduce the footprint of individual rule entries
which can be many in larger installations.

Change-Id: I866c79af9a5abe8719ead5b98e19a08c3cb7f647