7330 FIX Fix confusions when working with localized shell environments
When switching to the site user context we now enforce a non localized
environment. The reason for this configuration is that the parameters and
outputs of the monitoring plug-ins are localized. If they are called from the
core, they are always language-neutral. During manual testing, the plugins may
behave differently depending on the localization of the user's environment
variables. This can lead to confusion during tests.

CMK-1832

Change-Id: Ia7b677b0daa5f7945db64b835e6a58cc70c36ee4