Remove unused FILE_PATH global variable for .mk config files
Check_MK base config files could contain references to a global
variable FILE_PATH. This is an unused variable for a long time
now. It was replaced by FOLDER_PATH in the early days of WATO.

The FILE_PATH variable has now been cleaned up. If your
configuration files contain such a variable, you need to clean
it up or replace it with FOLDER_PATH.

Change-Id: I34f493ee5d8a3af1b762e10caa9b18cc9831b395