Cleaned up core specific autochecks loading
The CMC check helper specific autocheck file hierarchy was used when a
"world" parameter was set to "active". This parameter was propagated
several levels through the code to tell
cmk_base.autochecks.read_autochecks_of() to use a helper specific path.

Since each invokation of the Check_MK base code is either a helper or
not this can be swapped per execution in a central place.

This central place is the keepalive.enable() function now. It simply
changes the global path definitions in keepalive mode to point to
the keepalive helper specific directory.

CMK-1818

Change-Id: Ia58642b4e20a097aeb619daab52fb788b97a42f0