7385 FIX check_http: Fix port mixup in URL-mode
The check_http check in URL mode mixed up the servers port and the proxy port if a proxy was used.
If users were using a proxy to check a URL, the configured port was used as the proxys port, while
it was not possible to configure a port for the server. Users who have been relying on this wrong
behaviour need to edit their rules.
The rule has been reworked to clearly show which settings apply the server, and which to the proxy.

Change-Id: I2eef53f196032f8b718b3a7aad4c49832ba27191