RRD data query has configurable limit on data length
RRDtool has a configurable default limit of 400 points per
query. Livestatus did not configure that value until now, but we may need
to do queries over large periods of time and we don't want down sampling.

This change offers backwards compatibility, leaving this parameter as
optional.

The off by one in the query to result arises on the new way result bounds
are given. Now the result bounds contain the query bounds instead of
providing a shift of bounds between query and response.

CMK-1935

Change-Id: I2bcd22396887fe04f7617cb6e1248225662482c1