Show effective labels of hosts / service on WATO "Parameters of X" page
The page is used to show effective settings of the hosts and services,
most of them are based on ruleset configurations, but the pages are not
limited to this.

The page now also shows the effective labels for hosts and services
of all three possible sources (explicit host attribute, ruleset,
discovered).

Change-Id: I5d5e181b78a0c80b6e24a808504dbf7f684aa873