Fixed broken BI state filters
All filters are used by default since 1.6 in views just like it is
done in reports for a long time. Since then it is important that
the filters detect on their own whether or not they are active.

This is especially important for the filters that only use checkboxes
as values. In these filters all checkboxes can be unchecked which makes
the HTTP variables for this filter vanish. As a result the filter does
not know whether or not it has been used (with all checkboxes unchecked)
or it has not be used and should not filter.

To fix this problem there is now a hidden field which is always set when
the filter has been submitted with a form. This makes the filter know
whether or not to do it's work.

Also cleaned up the checkbox processing to use the generic
get_checkbox() logic.

CMK-1869

Change-Id: I0c2ceef836aa1e9ae9c4c115681020170973446d