Add specific visualization for labels depending on their source
Labels can be created explicit by host config, rulesets and can
be discovered automatically by Check_MK. Depending on the source
of a label it is now colorized in different ways to make the origin
of the label transparent to the user.

We may need to work a bit on the colors. That will be the difficult
part ;-).

CMK-1817

Change-Id: I4c52509408159ffd97c61150126f9ea94f1863c0