7338 FIX Dropped ancient "no_discovery_possible" from check API
The function no_discovery_possible was provided by the check API in the early days
were the check_info was declared as 3-element tuple. Since the checks are registered
as dictionaries this function is not needed anymore.

In case you get an issue with this after an update, you will have to migrate your
checks to the current check_info syntax.

Change-Id: Ied704a41cba892368f02f766af4c86999ddb3ab3