Thread an addition long output field from the monitoring log through to Livestatus.
This adds a "long_plugin_output" column to the "log" table and a
"long_log_output" column to the "statehist" table.  These columns are filled
from an additional field at the end of the following monitoring log entries:

   INITIAL HOST STATE
   INITIAL SERVICE STATE
   CURRENT HOST STATE
   CURRENT SERVICE STATE
   HOST ALERT
   SERVICE ALERT
   HOST NOTIFICATION
   SERVICE NOTIFICATION

Line breaks in the long output are represented as "\n" (i.e. a backslash
followed by the letter "n") in the monitoring log.  This is the same
encoding as is used for e.g. the UPDATE_SHADOW_HOST_STATE command.

CMK-1364

Change-Id: Idbaf12ed0d50611a9b9d70b6ad60f15310d3c4f4