mk_oracle: better support for mounted databases
The plugin {{mk_oracle}} showed multiple lines with errors when a database was not
opened. That created some problems in oracle_jobs, which have been fixed.
All other SQLs with select from the data dictionary have been redesigned to
reduce multiple error lines to one debug line.

Incomaptble Change:
This is an incompatble Change for installations with Databases who are not opened
in 'read write' mode. Databases in mounted state are not affected by this change!
The SQL for oracle_jobs selects data when the Database in Primary role and
opened 'READ WRITE'.

Please update the agent plugin {{mk_oracle}}

Change-Id: Ic3ba8895a2c2bd8fd1ad2a2c442b202cadc16db4