Separate CEE specific ConfigCache/HostConfig
The CEE specific configuration settings should only be available
when using CEE/CME and should be located below cmk_base.cee in
the module hierarchy.

Currently this can not be done easily. The best we can do at the
moment is to separate them into dedicated classes which can then
be moved over later.

CMK-2030

Change-Id: If7fe3973986c995af361e8929302860f1ab72c3b