oracle_dataguard_stats: Ignore switchoverstate in Primary without Broker
The switchoverstate in Primary Databases without a Standby-Database could lead
to wrong CRITs. The check ignore the switchoverstate in Primary when the
Broker is diabled.
Set the parameter primary_broker_state=True in environemnts with a Data-Guard
configuration and disabled Broker (init.ora dg_broker_start=false). Otherwise the
switchoverstate is ignred in the Primary-Database.

Change-Id: I300965cde2575c1f49ca19a1fb5d8345a80bda9b