7439 FIX PagerDuty notification when Host Check Command queries a service
PagerDuty notification plugin tried to parse perfdata and include it on the
Notification payload. This led to some buggy behavior on host problems when
using a Host check command. Since this information is not Processed by
PagerDuty only showed, we drop the perfdata information out of the
notification. Also for consistency as other notification plugins don't
include this information either.

FEED-3479

Change-Id: Iddb84b8e6b1000f9b2bfb136dbd84e54287468f7