7346 FIX Fix name conflict between builtin snmp tag group and aux tag snmp
The bultin tag group "snmp" for configuring the SNMP datasource of a host
and the related auxiliary tag "snmp" used the same ID. This resulted in
a name conflict between both which could cause problems e.g. in the rule
editor when using the auxiliary tag as condition.

The tag group has now been renamed to "snmp_ds" and the auxiliary tag ID
stays "snmp".

Change-Id: I3d96efa7b8122933731135f4e0e05194c5870637