Make tag group definitions available in base world
This is needed for working with the tag and tag group in the new
ruleset structure.

The new config file is a plain copy of the GUI tag group structure.  It
will never be read by WATO and only written out when the GUI tag config
file is written.

This makes some recently introduced hacks obsolete.

CMK-2031

Change-Id: I9d5e9c0cde53f8533f4eca23bd2617962550438d