7477 FIX job: Gracefully handle long running cronjobs
Previously mk_job got confused when a job was started while a job with
the same Identity was still running.

The mk_job script can now handle these cases. The corresponding check plugin
will apply the age levels to the longest runnging job, and additionally
display the stats of the last completed job.

Change-Id: I86e5d44bbd28e20ca8d353a0c6c37110bf6a530b