Add small helper tool cmk-update-config
This program will be executed automatically in these situations

a) During the "omd update" procedure
b) On remote sites after writing the received config when the sites of
the central and remote sites differ.

The purpose of this program is to prepare the Checkmk configuration of
the local site to be read by Checkmk. The configuration files might have
been written with another Checkmk version before which can not be read
into Checkmk without modification.

Currently the tool simply loads and saves the tag, ruleset and host- and
folder configuration every time the cmk-update-config is invoked. It
leaves the transformation of the data structures to the loading
and saving code of Checkmk.

In the future we may extend this program and add transformation mechanics
exclusively to this program to keep the Checkmk program code clean.
To be able to do this we would first need to add some additional featues
like version information to the individual config files.

Change-Id: Ie286927335eafcd73d6be0c82fb88e243af791d4