Removed the implicit "wato" tag
It was set automatically for all hosts that were configured using WATO.
It was not visible in WATO and not usable there. It could only be used
in hand crafted queries.

In case you need this in your scripts, you will have to create a tag
group in your configuration that produces such a tag.

Change-Id: Ifc950ab359a8265842b0ad37a28a122c9d19a6d8