7480 azure_usagedetails: Check Plugin to monitor Microsoft Azure costs
This check reports the daily costs of the Azure subscription,
in sum and per Azure service type and/or resource group.

Note that cost information has a granularity of one day.
The current value will always display the costs corresonding
to the last complete day (00:00 - 23:59 UTC).

Levels can be imposed on this value using the rule
"Azure Usage Details (Costs)".

One service per Azure service type plus an additional summary
are created on both the host running the agent and every
host corresponding to a monitored resource group.

CMK-2006

Change-Id: Ifb12d8fa1f6a6fb7cbfe07953c26bf19a7fc6e9a