7418 netapp_api_if: new option to disable a warning for interface groups
It is now possible to disable the warning "Interfaces do not have the
same speed" for interface groups monitored by the netapp_api_if
interface check. This can be configured in the rule "Network interfaces
and switch ports". Currently this option is only supported by the
netapp_api_if check.

Change-Id: Icb08646ba4bee9cf17fb1f8d3797dc4eee139e33