Update host tag rule condition editor
It's now possible to configure a list of tag IDs for a single tag
group. This reduces some situations where users had to configure
multiple rules when a rule should affect multiple choices of a
tag group.

In this step the tag condition input has been changed to be more
compact by default. This should improve the condition dialog usability,
especially for users with a larger number of tag groups.

CMK-2188

Change-Id: I9e6e8210d96c32938c4159406e444402a8b15b24