Prevent firefox from autocompleting hidden fields on page reload
The ListOfMultiple valuespec uses a hidden field to keep track of the
elements choosen by the user. On page reload the previously choosen
elements were reset to the initial value, but not the hidden field
value. This made the valuespec unusable until completely reopening the
page again.

This fix affects all all hidden fields with the assumption that this
logic is the correct one for all our dialogs.

Change-Id: I77e3e06c6384c3e07f914fd5a3906e8e9061e046