7482 FIX mk_mysql: More consistent naming of instances
The naming of different mysql instances has been simplified.
The item names of the monitored instances are now given by the corresponding sockets
(e.g. an instance using the socket "/var/run/mysql.sock" will be called "/var/run/mysql.sock").
The sockets are determined by the socket argument in the mysgl.cfg file or by the
<tt>--socket=</tt> option found in <tt>ps -fww -C mysqld</tt>.
If neiter of those is found, only one instance will be monitored, using the default name "mysql".

Previously as a third attempt the <tt>--user=</tt> option of the afore mentioned ps output has
been used as instance name. This is not longer considered.

Change-Id: Id2698b20c771913276c57f6eb475d88e70c7349e