7166 Agent Bakery now supports systemd
Until now, it was necessary to have xinetd installed on a host system
in order to setup the CheckMK Agent coming from the agent bakery
automatically.

The baked .rpm and .deb installation packages will now fall back
to install the Agent as a systemd-service, if no xinetd is available
on the host.

With this change, the baked agent package equals the vanilla agent
package in its installation behavior.

CMK-2066

Change-Id: I77f442467bcbe55faa4558eae2df3a94de71a1d1