7350 Make agent related rules available in bakery ruleset list
The rulesets "Encryption" and "TCP port for connection to Check_MK agent" are used
by Checkmk during checking but also by the Agent Bakery for configuring the baked
agent correctly. The rulesets are located in the "Access to Agents / Checkmk Agent"
ruleset group.

Users often try to find the ruleset in the wrong group while configuring the
Agent Bakery packages. For this reason the ruleset is now available in both groups.

Change-Id: Ic3fed0d69ce4bee162a5d1b2b71a5146dc17834c