7700 FIX mk_postgres: Fixed several problems
Originally the version of the client tools was taken, but what really matters
is the version of the server. Terminate the script, if the server can not be
queried.

In postgresql 9.2 the column state was introduced, which has to be queried to
find the state of the sessions. The column state can be NULL.

Here the order of the columns did not match with what the server side expects,
with the result, that there was never any active connection shown in the
web-gui.

Thanks to Felix E.

FEED-288

Change-Id: Id75e1453ec651ed1f42f211c39e783a757ae4c04